Move error-related data structures into a separate module
